refactor: 공통 도메인의 패키지명을 변경하라
