+++ module-common/src/main/java/kitchenpos/domain/Name.java
@@ -1,7 +1,7 @@
 package kitchenpos.domain;
 
-import kitchenpos.core.exception.BadRequestException;
-import kitchenpos.core.exception.ExceptionType;
+import kitchenpos.common.exception.BadRequestException;
+import kitchenpos.common.exception.ExceptionType;
 import org.springframework.util.StringUtils;
 
 import javax.persistence.Column;
